% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_dir_if_doesnt_exist}
\alias{create_dir_if_doesnt_exist}
\title{Creates a new directory if it doesn't already exist}
\usage{
create_dir_if_doesnt_exist(path)
}
\arguments{
\item{path}{The full path where the new directory should be created}
}
\value{
Nothing. This functions is called for its side effect.
}

