% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_fia.R
\name{download_fia}
\alias{download_fia}
\title{Download FIA tables.}
\usage{
download_fia(tables, states, destination_dir, dir_by_sate = TRUE,
  overwrite = FALSE)
}
\arguments{
\item{tables}{The names of the tables to be downloaded.}

\item{states}{TODO}

\item{destination_dir}{The directory where files will be saved to.}

\item{dir_by_sate}{Save tables from each state in a separate directory?}

\item{overwrite}{Should overwrite file if it already exists?}
}
\value{
Nothing. This function is called for its side effect.
}
\description{
Gets the Forest Inventory Scervice (FIA) data tables for each state
separatelly.
}
\details{
TODO
}

