% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_fia_tables.R
\name{download_fia_tables}
\alias{download_fia_tables}
\title{Downloads fia tables to dir}
\usage{
download_fia_tables(table_names, states, destination_dir, list_available_tables,
  table_format = "csv", dir_by_sate = TRUE, overwrite = TRUE)
}
\arguments{
\item{table_names}{Choose which tables to download}

\item{states}{Download data for which states. If "all", `data_tables_states_combined` will be downloaded}

\item{destination_dir}{Path to directory where the data files will be saved.}

\item{list_available_tables}{Object returned from `fia::list_available_tables`}

\item{table_format}{"csv" (default) or "zip"}

\item{dir_by_sate}{Save data for each state in a separate directory? Only works if states != `all` and for non-reference tables.}

\item{overwrite}{Should datasets be overwritten?}
}
\value{
Nothing. Side effect of saving files to a directory
}
\description{
Downloads fia tables to dir
}
