% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_available_tables.R
\name{list_available_tables}
\alias{list_available_tables}
\title{List available tables in the FIA dataset}
\usage{
list_available_tables(url = fia::fia_base_url())
}
\arguments{
\item{url}{The base url for the FIA database. It defaults to the url returned by `fia::fia_base_url()`}
}
\value{
TODO
}
\description{
Lists the available data tables from FIA.
}
\details{
TODO
}

