% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_available_tables.R
\name{list_available_tables}
\alias{list_available_tables}
\title{Lists tables in the FIA dataset}
\usage{
list_available_tables(url = fia::url_fia())
}
\arguments{
\item{url}{The base url for the FIA database. Defaults \code{fia::url_fia}}
}
\description{
Returns a list with the available FIA data.
}

